requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
spacy
textstat==0.7.3
sentence-transformers
torch
google-genai
pydantic>=2.0
pytest
pytest-asyncio
